+++ liquid-rust-driver/src/lib.rs
@@ -56,25 +56,4 @@ pub fn run_compiler(args: Vec<String>) -> i32 {
     })
 }
 
-fn test_file(file: &str, expected: liquid_rust_fixpoint::Safeness) -> bool { 
-    let s0:String = "target/debug/liquid-rust".to_string();
-    let s1:String = "--crate-type=lib".to_string();
-    let mut args: Vec<String> = vec!(s0, s1); 
-    args.push(file.to_string());
-    match run_compiler_result(args) {
-        Ok(r) => r.tag == expected,
-        _     => false,
-    }
-}
-
-/// Test a single file that should be SAFE
-pub fn test_safe(file: &str) -> bool {
-    test_file(file, liquid_rust_fixpoint::Safeness::Safe)
-}
-
-/// Test a single file that should be SAFE
-pub fn test_unsafe(file: &str) -> bool {
-    test_file(file, liquid_rust_fixpoint::Safeness::Unsafe)
-}
-
